Added files via upload
Finished! I know there is some polishing I could do and there are definitely best practices that I am missing, but I am pretty pleased with the outcome of about 15 or so working hours on this. When testing, make sure the images are stored in a file directory called "images".